Move ws-manager-mk2 CRD to API package
