+++ components/ws-manager-api/go/BUILD.yaml
@@ -10,4 +10,4 @@ packages:
       - "go.sum"
     config:
       packaging: library
-      dontTest: false
+      dontTest: true
