+++ components/ws-manager-api/go/crd/v1/groupversion_info.go
