+++ components/ws-manager-api/go/crd/v1/workspace_types.go
